Csv Enum BIG Data Decode. Still need to translate although on good track.
Also want to optimalize somehow this enum. I think of static class that will consist of other enums that will be seperate on the segments, cause it's flustrating to handle such big enums...

Morover DB soon be reorganize so we want to have better background prepare for easy handling :)

Cheers.